+++ Arduino/ODriveArduino/ODriveArduino.h
@@ -26,6 +26,7 @@ public:
     void SetPosition(int motor_number, float position, float velocity_feedforward, float current_feedforward);
     void SetVelocity(int motor_number, float velocity);
     void SetVelocity(int motor_number, float velocity, float current_feedforward);
+    void SetCurrent(int motor_number, float current);
 
     // General params
     float readFloat();