+++ Firmware/communication/communication.cpp
@@ -91,12 +91,8 @@ void init_communication(void) {
         osDelay(1);
 }
 
-
 float oscilloscope[OSCILLOSCOPE_SIZE] = {0};
 size_t oscilloscope_pos = 0;
-
-
-static CAN_context can1_ctx;
 
 // Helper class because the protocol library doesn't yet
 // support non-member functions
@@ -136,6 +132,7 @@ static inline auto make_obj_tree() {
             make_protocol_ro_property("min_stack_space_comms", &system_stats_.min_stack_space_comms),
             make_protocol_ro_property("min_stack_space_usb", &system_stats_.min_stack_space_usb),
             make_protocol_ro_property("min_stack_space_uart", &system_stats_.min_stack_space_uart),
+            make_protocol_ro_property("min_stack_space_can", &system_stats_.min_stack_space_can),
             make_protocol_ro_property("min_stack_space_usb_irq", &system_stats_.min_stack_space_usb_irq),
             make_protocol_ro_property("min_stack_space_startup", &system_stats_.min_stack_space_startup),
             make_protocol_object("usb",
@@ -163,11 +160,14 @@ static inline auto make_obj_tree() {
             make_protocol_object("gpio2_pwm_mapping", make_protocol_definitions(board_config.pwm_mappings[1])),
             make_protocol_object("gpio3_pwm_mapping", make_protocol_definitions(board_config.pwm_mappings[2])),
 #endif
-            make_protocol_object("gpio4_pwm_mapping", make_protocol_definitions(board_config.pwm_mappings[3]))
-        ),
+            make_protocol_object("gpio4_pwm_mapping", make_protocol_definitions(board_config.pwm_mappings[3])),
+
+            make_protocol_object("gpio3_analog_mapping", make_protocol_definitions(board_config.analog_mappings[2])),
+            make_protocol_object("gpio4_analog_mapping", make_protocol_definitions(board_config.analog_mappings[3]))
+            ),
         make_protocol_object("axis0", axes[0]->make_protocol_definitions()),
         make_protocol_object("axis1", axes[1]->make_protocol_definitions()),
-        make_protocol_object("can", can1_ctx.make_protocol_definitions()),
+        make_protocol_object("can", odCAN->make_protocol_definitions()),
         make_protocol_property("test_property", &test_property),
         make_protocol_function("test_function", static_functions, &StaticFunctions::test_function, "delta"),
         make_protocol_function("get_oscilloscope_val", static_functions, &StaticFunctions::get_oscilloscope_val, "index"),
@@ -202,8 +202,7 @@ void communication_task(void * ctx) {
     if (board_config.enable_i2c_instead_of_can) {
         start_i2c_server();
     } else {
-        // TODO: finish implementing CAN
-        // start_can_server(can1_ctx, CAN1, serial_number);
+        odCAN->start_can_server();
     }
 
     for (;;) {